requestbuyitem weight fix by striker
git-svn-id: http://svn.l2jfree.com/l2jfree@4750 6ff00ba9-25cd-4355-987e-099c62bc2628

+++ trunk/l2jfree-core/src/main/java/com/l2jfree/gameserver/network/clientpackets/RequestBuyItem.java
@@ -272,7 +272,7 @@ public class RequestBuyItem extends L2GameClientPacket
 			taxedPriceTotal += taxedPrice;
 			taxTotal += tax;
 
-			weight += (count * template.getWeight());
+			weight += (long) count * template.getWeight();
 			if (!template.isStackable())
 				slots += count;
 			else if (player.getInventory().getItemByItemId(itemId) == null)
